backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# AI / LangChain
langchain==0.3.14
langchain-google-genai==2.0.8
langchain-groq==0.2.4
langchain-community==0.3.14

# Vector search
faiss-cpu==1.9.0.post1

# Data sources
yfinance==0.2.51
requests==2.32.3

# Utilities
pydantic[email]==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.20
httpx==0.28.1
